Separate Avro and ProtoBuf things in their own adapter packages
