+++ adapter/avro/mu-avro.cabal
@@ -1,0 +1,47 @@
+cabal-version:       >=1.10
+name:                mu-avro
+version:             0.1.0.0
+synopsis:            Avro serialization support for Mu microservices
+-- description:
+-- bug-reports:
+license:             Apache-2.0
+license-file:        LICENSE
+author:              Alejandro Serrano, Flavio Corpa
+maintainer:          alejandro.serrano@47deg.com
+-- copyright:
+category:            Network
+build-type:          Simple
+
+library
+  exposed-modules:     Mu.Adapter.Avro
+                     , Mu.Adapter.Avro.Example
+                     , Mu.Quasi.Avro
+  -- other-modules:
+  -- other-extensions:
+  build-depends:       base >=4.12 && <5
+                     , mu-schema
+                     , avro
+                     , tagged
+                     , aeson
+                     , text
+                     , vector
+                     , containers
+                     , unordered-containers
+                     , sop-core
+                     , bytestring
+                     , template-haskell >= 2.12
+  hs-source-dirs:      src
+  default-language:    Haskell2010
+  ghc-options:         -Wall
+                       -fprint-potential-instances
+
+executable test-avro
+  main-is:             Avro.hs
+  build-depends:       base >=4.12 && <5
+                     , mu-schema
+                     , mu-avro
+                     , avro
+                     , bytestring
+  hs-source-dirs:      test
+  default-language:    Haskell2010
+  ghc-options:         -Wall
