+++ adapter/avro/test/avro/consume.py
