+++ adapter/avro/test/avro/example.avsc
