+++ adapter/avro/test/avro/generate.py
