+++ adapter/protobuf/test/protobuf/consume.py
