+++ adapter/protobuf/test/protobuf/example.proto
