+++ adapter/protobuf/test/protobuf/generate.py
