+++ examples/health-check/src/ClientRecord.hs
@@ -10,7 +10,7 @@ import qualified Data.Text as T
 import GHC.Generics (Generic)
 import System.Environment
 
-import Mu.Client.GRpc.Record
+import Mu.GRpc.Client.Record
 
 import Definition
