+++ grpc/mu-grpc.cabal
@@ -18,38 +18,45 @@ build-type:          Simple
 extra-source-files:  CHANGELOG.md
 
 library
-  exposed-modules:     Mu.Server.GRpc,
-                       Mu.Client.GRpc.Internal,
-                       Mu.Client.GRpc.TyApps,
-                       Mu.Client.GRpc.Record,
-                       Mu.Client.GRpc.Examples
-  other-modules:       Mu.GRpc.Shared
+  exposed-modules:     Mu.GRpc.Quasi,
+                       Mu.GRpc.Server,
+                       Mu.GRpc.Client.TyApps,
+                       Mu.GRpc.Client.Record,
+                       Mu.GRpc.Client.Examples
+  other-modules:       Mu.GRpc.Shared,
+                       Mu.GRpc.Client.Internal
   -- other-extensions:
   build-depends:       base >=4.12 && <5, sop-core,
-                       mu-schema, mu-rpc, warp-grpc,
+                       mu-schema, mu-rpc, mu-protobuf,
+                       warp-grpc,
                        conduit, bytestring, text,
                        wai, warp, warp-tls,
                        async, stm, stm-conduit, stm-chans,
                        http2, http2-client,
                        http2-grpc-types, http2-client-grpc,
                        proto3-wire, http2-grpc-proto3-wire,
-                       template-haskell, th-abstraction
+                       template-haskell, th-abstraction,
+                       language-protobuf, compendium-client,
+                       http-client, servant-client-core
   hs-source-dirs:      src
   default-language:    Haskell2010
   ghc-options:         -Wall -fprint-potential-instances
 
 executable grpc-example-server
   main-is:             ExampleServer.hs
+  other-modules:       Mu.GRpc.Server, Mu.GRpc.Shared
   build-depends:       base >=4.12 && <5, sop-core,
-                       mu-schema, mu-rpc, warp-grpc,
+                       mu-schema, mu-rpc, mu-protobuf,
+                       warp-grpc,
                        conduit, bytestring, text,
                        wai, warp, warp-tls,
                        async, stm, stm-conduit, stm-chans,
                        http2, http2-client,
                        http2-grpc-types, http2-client-grpc,
                        proto3-wire, http2-grpc-proto3-wire,
-                       template-haskell, th-abstraction
-  other-modules:       Mu.GRpc.Shared, Mu.Server.GRpc
+                       template-haskell, th-abstraction,
+                       language-protobuf, compendium-client,
+                       http-client, servant-client-core
   hs-source-dirs:      src
   default-language:    Haskell2010
   ghc-options:         -Wall
