+++ rpc/mu-rpc.cabal
@@ -19,16 +19,13 @@ extra-source-files:  README.md, CHANGELOG.md
 
 library
   exposed-modules:     Mu.Rpc,
-                       Mu.Rpc.Quasi,
                        Mu.Server,
                        Mu.Rpc.Examples
   -- other-modules:
   -- other-extensions:
   build-depends:       base >=4.12 && <5, sop-core,
                        mu-schema, conduit, text,
-                       template-haskell, language-protobuf,
-                       compendium-client,
-                       http-client, servant-client-core
+                       template-haskell
   hs-source-dirs:      src
   default-language:    Haskell2010
   ghc-options:         -Wall -fprint-potential-instances
