+++ schema/mu-schema.cabal
@@ -25,53 +25,24 @@ library
                      , Mu.Schema.Interpretation.Anonymous
                      , Mu.Schema.Class
                      , Mu.Schema.Registry
-                     , Mu.Schema.Adapter.Avro
-                     , Mu.Schema.Adapter.ProtoBuf
-                     , Mu.Schema.Adapter.Json
-                     , Mu.Schema.Quasi
                      , Mu.Schema.Conversion.TypesToSchema
                      , Mu.Schema.Conversion.SchemaToTypes
                      , Mu.Schema.Examples
-                     , Mu.Schema.AvroExample
+                     , Mu.Schema.Annotations
+                     , Mu.Adapter.Json
   -- other-modules:
   -- other-extensions:
   build-depends:       base >=4.12 && <5
                      , sop-core
                      , containers
                      , unordered-containers
-                     , vector
                      , bytestring
+                     , vector
                      , text
-                     , avro
-                     , tagged
-                     , proto3-wire
                      , aeson
                      , template-haskell >= 2.12
                      , th-abstraction
-                     , language-protobuf
   hs-source-dirs:      src
   default-language:    Haskell2010
   ghc-options:         -Wall
                        -fprint-potential-instances
-
-executable test-avro
-  main-is:             Avro.hs
-  build-depends:       base >=4.12 && <5
-                     , sop-core
-                     , mu-schema
-                     , avro
-                     , bytestring
-  hs-source-dirs:      test
-  default-language:    Haskell2010
-  ghc-options:         -Wall
-
-executable test-protobuf
-  main-is:             ProtoBuf.hs
-  build-depends:       base >=4.12 && <5
-                     , sop-core
-                     , mu-schema
-                     , proto3-wire
-                     , bytestring
-  hs-source-dirs:      test
-  default-language:    Haskell2010
-  ghc-options:         -Wall
