+++ schema/src/Mu/Schema/Annotations.hs
@@ -1,0 +1,9 @@
+{-# language DataKinds, KindSignatures #-}
+module Mu.Schema.Annotations where
+
+import GHC.TypeLits
+
+-- ANNOTATION FOR CONVERSION
+
+data ProtoBufId (n :: Nat)
+data ProtoBufOneOfIds (ns :: [Nat])
