+++ schema/src/Mu/Schema/Examples.hs
@@ -8,19 +8,12 @@
 module Mu.Schema.Examples where
 
 import qualified Data.Aeson as J
-import qualified Data.Avro as A
 import qualified Data.Text as T
 import GHC.Generics
 
 import Mu.Schema
-import Mu.Schema.Adapter.Avro ()
-import Mu.Schema.Adapter.ProtoBuf
-import Mu.Schema.Adapter.Json ()
-
 import Mu.Schema.Conversion.SchemaToTypes
-
-import qualified Proto3.Wire.Encode as PBEnc
-import qualified Proto3.Wire.Decode as PBDec
+import Mu.Adapter.Json ()
 
 data Person
   = Person { firstName :: T.Text
@@ -30,26 +23,20 @@ data Person
            , address   :: Address }
   deriving (Eq, Show, Generic)
   deriving (HasSchema ExampleSchema "person")
-  deriving (A.HasAvroSchema, A.FromAvro, A.ToAvro, J.ToJSON, J.FromJSON)
+  deriving (J.ToJSON, J.FromJSON)
     via (WithSchema ExampleSchema "person" Person)
-
-personToProtoBuf :: Person -> PBEnc.MessageBuilder
-personToProtoBuf = toProtoViaSchema @ExampleSchema
-
-protoBufToPerson :: PBDec.Parser PBDec.RawMessage Person
-protoBufToPerson = fromProtoViaSchema @ExampleSchema
 
 data Address
   = Address { postcode :: T.Text
             , country  :: T.Text }
   deriving (Eq, Show, Generic)
   deriving (HasSchema ExampleSchema "address")
-  deriving (A.HasAvroSchema, A.FromAvro, A.ToAvro, J.ToJSON, J.FromJSON)
+  deriving (J.ToJSON, J.FromJSON)
     via (WithSchema ExampleSchema "address" Address)
 
 data Gender = Male | Female | NonBinary
   deriving (Eq, Show, Generic)
-  deriving (A.HasAvroSchema, A.FromAvro, A.ToAvro, J.ToJSON, J.FromJSON)
+  deriving (J.ToJSON, J.FromJSON)
     via (WithSchema ExampleSchema "gender" Gender)
 
 -- Schema for these data types
@@ -57,7 +44,7 @@ type ExampleSchema
   = '[ 'DEnum   "gender" '[]
                '[ 'ChoiceDef "male"   '[ ProtoBufId 1 ]
                 , 'ChoiceDef "female" '[ ProtoBufId 2 ]
-                , 'ChoiceDef "nb"     '[ ProtoBufId 0 ] ]
+                , 'ChoiceDef "nb"     '[ ProtoBufId 3 ] ]
      , 'DRecord "address" '[]
                '[ 'FieldDef "postcode" '[ ProtoBufId 1 ] ('TPrimitive T.Text)
                 , 'FieldDef "country"  '[ ProtoBufId 2 ] ('TPrimitive T.Text) ]
@@ -90,7 +77,7 @@ type ExampleSchema2
   = '[ 'DEnum   "gender" '[]
                '[ 'ChoiceDef "Male"      '[ ProtoBufId 1 ]
                 , 'ChoiceDef "Female"    '[ ProtoBufId 2 ]
-                , 'ChoiceDef "NonBinary" '[ ProtoBufId 0 ] ]
+                , 'ChoiceDef "NonBinary" '[ ProtoBufId 3 ] ]
      , 'DRecord "address" '[]
                '[ 'FieldDef "postcode" '[ ProtoBufId 1 ] ('TPrimitive T.Text)
                 , 'FieldDef "country"  '[ ProtoBufId 2 ] ('TPrimitive T.Text) ]
@@ -104,16 +91,3 @@ type ExampleSchema2
 
 type ExampleRegistry
   = '[ 2 ':-> ExampleSchema2, 1 ':-> ExampleSchema]
-
-type ExampleSchema3 = [protobuf|
-enum gender {
-  male      = 1;
-  female    = 2;
-  nonbinary = 3;
-}
-message person {
-  repeated string names = 1;
-  int age = 2;
-  gender gender = 3;
-}
-|]
