+++ stack.yaml
@@ -1,7 +1,9 @@
-resolver: lts-14.13
+resolver: lts-14.14
 
 packages:
 - schema
+- adapter/avro
+- adapter/protobuf
 - rpc
 - grpc
 - examples/health-check
